entrypoint/requirements.txt
kafka-python==1.4.4
Flask==0.12.2
avro-python3